corrected background and foreground handling
they would be set to "initial" which isn't right
if you've specified them. It will now use the foreground
and background colors determined in oho.cr
AND fall back to "initial" when it doesn't know what else to
do.

This involved changing the options hash passed to the Converter
and then passing that on to the EscapeCode initializer